requirements.txt
flask
pyrebase4
opencv-python
torch
transformers
pillow
numpy
gunicorn